% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createBerryData}
\alias{createBerryData}
\title{Utility function to create the data on page 63 of Berry et al}
\usage{
createBerryData(binary = FALSE)
}
\arguments{
\item{binary}{if \code{FALSE} return data at the study level.  Otherwise,
return data at the subject level}
}
\value{
if \code{binary==FALSE}, a tibble with columns Study, Subjects and
 Events.  Otherwise, a tibble containing columns Study, SubjectID and Event
}
\description{
Utility function to create the data on page 63 of Berry et al
}
\examples{
data <- createBerryData()
}
