% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createQtlPlot}
\alias{createQtlPlot}
\title{Create the QTL plot}
\usage{
createQtlPlot(mcmcData, mcmcVar = Value, groupData,
  groupResponse = ObservedResponse, groupSize = N, groupType = NULL,
  basicTheme = ggplot2::theme_light, qtls = c(0.1, 0.2, 0.8, 0.9),
  postScaleFactor = NULL, mcmcAlpha = 0.2, nDensity = 512)
}
\arguments{
\item{mcmcData}{the tibble containing the posterior MCMC samples of the metric}

\item{mcmcVar}{the column in \code{mcmcData} containing the sampled values of
the metric}

\item{groupData}{the tibble containing the observed group data}

\item{groupResponse}{the column in \code{groupData} containing the observed
by-group metric values}

\item{groupSize}{the column in the \code{groupData} containing the group sizes}

\item{groupType}{The column in \code{groupData} containing the group type for
for the observation.  If \code{NULL}, group types are not identified in the
plot.  Otherwise, \code{groupType} is converted to a factor and used as an
aesthetic.}

\item{basicTheme}{one of the themes provided by \code{ggplot2}, for example
\code{theme_light} or \code{theme_dark}.}

\item{qtls}{a list containing the quantiles that define the QTL thresholds.
If \code{NULL} or of zero length, the QTL thresholds are not shown.}

\item{postScaleFactor}{scales the posterior density relative to the height of 
the \code{groupSize} bars.  If \code{NULL}, \code{postScaleFactor} is chosen 
so that the mode of the posterior is twice as high as the tallest group size 
bar.  See Usage Notes below.}

\item{mcmcAlpha}{the transparency of the fill used to indicate QTL bands}

\item{nDensity}{the number of bins used to estimate the density.  Must be a 
power of two.  Normally, this can be left unchanged, but occasionally a
different value is needed to correct visual artefacts related to the shading
of the plot}
}
\value{
the QTL plot
}
\description{
Create the QTL plot
}
\section{Usage Notes}{

The plot returned by the function is deliberately crude: the x axis is 
unlabelled, the legends are unfomatted, and so on.  However, it can be easily 
customised to include more appropriate titles, legends and so forth, as shown
in the second example.

The scaled kernel estimate of the density is multiplied by
\code{postScaleFactor} to obtain the curve shown in the plot.  The height of 
the scaled kernel density at its mode is 1.  Therefore, by default, 
\code{postScaleFactor} is set to twice the largest group size.
}

\examples{
b <- createBerryData() \%>\% 
       tibble::add_column(Region=as.factor(c(1,1,1,2,2,2,3,3,3)))
m <- fitBinomialModel(b$Subjects, b$Events) \%>\% dplyr::filter(Index==10)
createQtlPlot(groupData=b,
              groupResponse=ObservedResponse,
              groupSize=Subjects,
              groupType=Region,
              mcmcData=m,
              mcmcVar=Value) +
              ggplot2::scale_fill_manual(name="QTL",
                                         labels=c("Action low",
                                                  "Warn low",
                                                  "",
                                                  "Warn high",
                                                  "Action high"),
                                         values=c("red",
                                                  "yellow",
                                                  "white",
                                                  "yellow",
                                                  "red")) +
              ggplot2::scale_colour_manual(name="Region",
                                           values=1:3,
                                           labels=c("US",
                                                    "EMEA",
                                                    "Other")) +
              ggplot2::labs(x="Observed response",
                            title="Example QTL plot")
}
