% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitTteModel}
\alias{fitTteModel}
\title{Fit a time-to-event QTL model}
\usage{
fitTteModel(time, status, censorFlag = 0, g, inits = NULL,
  modelString = NULL, autoRun = TRUE, raw = FALSE, ...)
}
\arguments{
\item{time}{a vector containing the event/censoring times for each datum.}

\item{status}{a vector containing the censoring status for each datum.}

\item{censorFlag}{a scalar or vector containing the values of \code{status}
that indicate that the corresponding event time is censored.}

\item{g}{a vector indicating the group membership for each datum.  Either a 
factor defining group membership, or a vector of integers between 1 and the
number of groups.}

\item{inits}{a list containing the initial values of the hyperparameters.
See Usage Notes below.}

\item{modelString}{The JAGS model string that defines the model to be fitted}

\item{autoRun}{Logical.  If \code{TRUE}, use \code{autorun.jags}.  Otherwise,
use \code{run.jags}.}

\item{raw}{logical if \code{TRUE}, the \code{mcmc} object created by
 \code{run.jags()} is returned.  Otherwise, a \code{tibble} containing the
concatenation of elements of the \code{mcmc.list} created by JAGS
is returned}

\item{...}{passed to JAGS}
}
\value{
Either the \code{mcmc} (or \code{mcmc.list}) object returned by
JAGS or a tibble containing the MCMC samples from the posterior
distribution
}
\description{
Fit a time-to-event QTL model
}
\section{Usage notes}{

If \code{modelString == NULL}, the model string is obtained by calling
\code{getModelString("tte")}.\cr\cr  
If \code{raw == FALSE}, the chain from which each observation is drawn is
indicated by \code{Chain} and the dataset is transformed into tidy format,
with the model parameter indicated by \code{Parameter}.\cr\cr  
The \code{inits} parameter can be used to define the number of chains created
my JAGS.  If a list of lists, the number of elements in the
outer list defines the number of chains and the elements of each sub-list
define the initial value for each hyperparameter.  For example, the default
value of \code{inits} requests two chains.  The initial values of \code{shape}
 and \code{scale} in the first chain are both 0.0001.  In the second
chain, they are both 1.  If the MCMC model has converged and is stationary, 
the initial values of the hyperparameters will be irrelevant.   To check for 
convergence, it is necessary - but not sufficient - to obtain more than one
chain and to use different initial values for each chain.\cr\cr
Note that some parameters to \code{run.jags} are incompatible with 
\code{autorun.jags} and will therefor cause an error when passed in 
\code{...} unless \code{autoRun} is set to \code{FALSE}.  These parameters
include \code{samples} and \code{burnin} and will result in an error similar
to "Error in \code{extend.jags(runjags.object, add.monitor = add.monitor,
drop.monitor = drop.monitor,}  : formal argument "burnin" matched by multiple
actual arguments".
}

\examples{
#Fit a QTL model to the VA lung dataset reported by Kalbfleisch and 
#Prentice (1980), using cell type to define group membership
fitTteModel(time=vaLung$SurvivalTime,
            status=vaLung$Status,
            g=vaLung$CellType)
}
\seealso{
\code{\link{fitBinomialModel}}, \code{\link{fitPoissonModel}},
\code{\link{fitBinaryModel}}
}
