% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getModelString}
\alias{getModelString}
\title{Function to obtain the string that defines the default JAGS model for each
data type}
\usage{
getModelString(label = c("binary", "binomial", "normal", "poisson",
  "tte"))
}
\arguments{
\item{label}{the type of model}
}
\value{
a string defining the required model
}
\description{
Function to obtain the string that defines the default JAGS model for each
data type
}
\section{Usage notes}{

\code{label} is determined case insensitively
}

\examples{
#To fit a binary model
binModel <- getModelString("binary")
}
