% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qtlFromQuantile}
\alias{qtlFromQuantile}
\title{Obtain the QTL based on the quantile of a sampled posterior}
\usage{
qtlFromQuantile(data, probs, type = 4)
}
\arguments{
\item{data}{a tibble containing the MCMC samples from which the quantile
should be estimated}

\item{probs}{a vector containing the required quantiles}

\item{type}{the quantile type to be returned.  See the help for \code{quantile()} for
more information.  Default 4, to ensure quality with sampled values.}
}
\value{
a vector containing the quantiles obtained from the Value column in 
\code{data}
}
\description{
Obtain the quantile based on a sampled posterior
}
\examples{
b <- createBerryData()
m <- fitBinomialModel(b$Subjects, b$Events) \%>\% dplyr::filter(Index==10)
qtlFromQuantile(m, c(0.1, 0.2, 0.8, 0.9))
}
