% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runjagsResultsToTibble}
\alias{runjagsResultsToTibble}
\title{A utility function to convert a \code{runjags} or \code{runjags.list} object
to a tibble}
\usage{
runjagsResultsToTibble(obj, paramPrefix, ...)
}
\arguments{
\item{obj}{the object to be converted}

\item{paramPrefix}{the name of any indexed model parameters, including the
opening square bracket. eg 'prob['}

\item{...}{arbitrary arguments that were passed to runjags when \code{obj}
was created}
}
\value{
the corresponding tibble
}
\description{
A utility function to convert a \code{runjags} or \code{runjags.list} object
to a tibble
}
